Move DOMContext to an interface (no implementation changes)
