+++ packages/tempots-ui/package.lib.json
@@ -1,6 +1,6 @@
 {
   "name": "@tempots/ui",
-  "version": "0.22.0",
+  "version": "0.22.1",
   "type": "module",
   "main": "./index.cjs",
   "module": "./index.js",
